# openapi.json
{
  "components": {
    "schemas": {
      "ArrayOutput": {
        "properties": {
          "data_b64": {
            "title": "Data B64",
            "type": "string"
          },
          "dtype": {
            "default": "float32",
            "title": "Dtype",
            "type": "string"
          },
          "kind": {
            "const": "array",
            "title": "Kind",
            "type": "string"
          },
          "shape": {
            "items": {
              "type": "integer"
            },
            "title": "Shape",
            "type": "array"
          },
          "unit": {
            "default": "",
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "shape",
          "data_b64"
        ],
        "title": "ArrayOutput",
        "type": "object"
      },
      "ConfigResponse": {
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/InputInfo"
            },
            "title": "Inputs",
            "type": "array"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "outputs": {
            "items": {
              "$ref": "#/components/schemas/OutputInfo"
            },
            "title": "Outputs",
            "type": "array"
          },
          "screens": {
            "items": {
              "$ref": "#/components/schemas/ScreenInfo"
            },
            "title": "Screens",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "model",
          "version",
          "inputs",
          "outputs",
          "screens"
        ],
        "title": "ConfigResponse",
        "type": "object"
      },
      "EvaluateV1Request": {
        "properties": {
          "inputs": {
            "additionalProperties": {
              "type": "number"
            },
            "default": {},
            "title": "Inputs",
            "type": "object"
          },
          "max_particles": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Particles"
          },
          "outputs": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Outputs",
            "type": "array"
          },
          "screen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Screen"
          },
          "smooth_images_sigma_px": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Smooth Images Sigma Px"
          }
        },
        "title": "EvaluateV1Request",
        "type": "object"
      },
      "EvaluateV1Response": {
        "properties": {
          "frame_index": {
            "title": "Frame Index",
            "type": "integer"
          },
          "inputs": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Inputs",
            "type": "object"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "outputs": {
            "additionalProperties": {
              "discriminator": {
                "mapping": {
                  "array": "#/components/schemas/ArrayOutput",
                  "particles": "#/components/schemas/ParticlesOutput",
                  "scalar": "#/components/schemas/ScalarOutput",
                  "value": "#/components/schemas/ValueOutput"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ScalarOutput"
                },
                {
                  "$ref": "#/components/schemas/ArrayOutput"
                },
                {
                  "$ref": "#/components/schemas/ParticlesOutput"
                },
                {
                  "$ref": "#/components/schemas/ValueOutput"
                }
              ]
            },
            "title": "Outputs",
            "type": "object"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "model",
          "version",
          "timestamp",
          "frame_index",
          "inputs",
          "outputs"
        ],
        "title": "EvaluateV1Response",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InputInfo": {
        "description": "A writable scalar knob.",
        "properties": {
          "constant": {
            "default": false,
            "title": "Constant",
            "type": "boolean"
          },
          "default": {
            "title": "Default",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "max": {
            "title": "Max",
            "type": "number"
          },
          "min": {
            "title": "Min",
            "type": "number"
          },
          "range_source": {
            "title": "Range Source",
            "type": "string"
          },
          "unit": {
            "default": "",
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "id",
          "default",
          "min",
          "max",
          "range_source"
        ],
        "title": "InputInfo",
        "type": "object"
      },
...
      "OutputInfo": {
        "properties": {
          "element_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Element Name"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "scalar",
              "array",
              "particles",
              "value"
            ],
            "title": "Kind",
            "type": "string"
          },
          "shape": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shape"
          },
          "unit": {
            "default": "",
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind"
        ],
        "title": "OutputInfo",
        "type": "object"
      },
      "ParticlesOutput": {
        "properties": {
          "coords": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Coords",
            "type": "object"
          },
          "kind": {
            "const": "particles",
            "title": "Kind",
            "type": "string"
          },
          "n": {
            "title": "N",
            "type": "integer"
          },
          "stats": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Stats",
            "type": "object"
          },
          "stats_units": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Stats Units",
            "type": "object"
          },
          "units": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Units",
            "type": "object"
          }
        },
        "required": [
          "kind",
          "n",
          "units",
          "coords",
          "stats",
          "stats_units"
        ],
        "title": "ParticlesOutput",
        "type": "object"
      },
...
      "ScreenInfo": {
        "description": "A diagnostic location. `image` is null when the model publishes no image for it.",
        "properties": {
          "image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "particles": {
            "title": "Particles",
            "type": "string"
          }
        },
        "required": [
          "key",
          "particles"
        ],
        "title": "ScreenInfo",
        "type": "object"
      },
      "SnapshotResponse": {
        "properties": {
          "inputs": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Inputs",
            "type": "object"
          }
        },
        "required": [
          "inputs"
        ],
        "title": "SnapshotResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "LUME model API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/models/{name}/config": {
      "get": {
        "description": "Everything a client needs to drive this model: inputs, outputs and screens.\n\nDerived from the model instance itself, so it is accurate for whatever `LUME_MODELS`\nselected. `range_source` says whether an input's min/max came from the model or was\nderived from its default, and `constant` marks an input the model pins to one value.",
        "operationId": "get_config_api_v1_models__name__config_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Everything a client needs to drive one model",
        "tags": [
          "model-api"
        ]
      }
    },
    "/api/v1/models/{name}/evaluate": {
      "post": {
        "description": "Stateless model evaluation. The one evaluate endpoint for every caller.\n\n`inputs` maps input id to control value, overlaid on the model's baseline, so send only\nthe knobs you want to change (`{}` is the design machine). The effective post-merge\nvalues come back in the response's `inputs`.\n\nAsk for outputs by id in `outputs`, or pass `screen` to get that screen's particle\ndistribution plus its image when it has one. Both may be combined. Every requested id is\npresent in `outputs`, keyed by id, and each value declares its own `kind` and unit.\n\nLarge arrays are base64-encoded little-endian float32, so decode with e.g.\n`numpy.frombuffer(base64.b64decode(s), dtype='<f4')` and reshape to `shape`. Particle\ncoordinates are subsampled to `max_particles` (default 3000, never the full beam).\n\nValues are in the model's own units, declared in the payload. Nothing is converted\nserver-side, so do not hard-code units.",
        "operationId": "evaluate_v1_api_v1_models__name__evaluate_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateV1Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluateV1Response"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run one model on a set of inputs and return the requested outputs",
        "tags": [
          "model-api"
        ]
      }
    },
...
    "/api/v1/models/{name}/machine-snapshot": {
      "get": {
        "description": "The live input values, read-only. Ids with no live value report their baseline.",
        "operationId": "machine_snapshot_api_v1_models__name__machine_snapshot_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The live input values for one model, read-only",
        "tags": [
          "model-api"
        ]
      }
    },
...
  }
}